handle mappingProductPrice, cartChecked, productChecked
